refactor(generator): split DataRecords and DtoFinder below 300-line cap
Move mutable working types (ModuleInfo, EndpointInfo, ViewInfo, etc.) from
DataRecords.cs into WorkingTypes.cs, and extract ExtractDtoProperties +
HasJsonIgnoreAttribute from DtoFinder.cs into DtoPropertyExtractor.cs.
